Opc Permissions

+++ libs/access/config/sql/mysql/access_role_add.sql
@@ -1,7 +1,7 @@
 drop procedure if exists access_role_add;
 go
 
-create procedure access_role_add( _role_id int unsigned, _allowed tinyint unsigned, _denied tinyint unsigned, _resourceTarget varchar(64), _schema varchar(256), _resourceName varchar(64), _criteria varchar(832), out _permission_id int unsigned )
+create procedure access_role_add( _role_id int unsigned, _allowed bigint unsigned, _denied bigint unsigned, _resourceTarget varchar(64), _schema varchar(256), _resourceName varchar(64), _criteria varchar(832), out _permission_id int unsigned )
 begin
 	declare _resource_id smallint unsigned;
 	select resource_id
